requirements.txt
httpx>=0.27
anthropic>=0.54
google-genai>=1.30
feedparser>=6.0
